moved int<->float conversion from sounddevices to Resampler
git-svn-id: file:///openmsx-svn/code/openmsx/trunk@6481 7512d703-071c-0410-b5c2-d2bdd895b7d1

+++ src/sound/Resample.hh
@@ -15,15 +15,15 @@ protected:
 	virtual ~Resample();
 
 	void setResampleRatio(double inFreq, double outFreq);
-	bool generateOutput(float* dataOut, unsigned num);
-	virtual bool generateInput(float* buffer, unsigned num) = 0;
+	bool generateOutput(int* dataOut, unsigned num);
+	virtual bool generateInput(int* buffer, unsigned num) = 0;
 
 private:
 	typedef FixedPoint<16> FilterIndex;
 	static const unsigned BUF_LEN = 16384;
 
-	void calcOutput(FilterIndex startFilterIndex, float* output);
-	void calcOutput2(float lastPos, float* output);
+	void calcOutput(FilterIndex startFilterIndex, int* output);
+	void calcOutput2(float lastPos, int* output);
 	void prepareData(unsigned extra);
 
 	float ratio;
